requirements.txt
fastapi==0.115.12
sqlmodel==0.0.24
PyMySQL==1.1.1
cryptography==44.0.2
uvicorn==0.34.0
python-dotenv==1.1.0
phpserialize==1.3